architect: record cpp mutation blockers
